# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "江都 API",
    "version": "1.1.0",
    "description": "小程序+管理后台共用接口api"
  },
  "host": "localhost:4000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/api/policys/": {
      "post": {
        "tags": [
          "Policy"
        ],
        "summary": "获取policy",
        "description": "获取政策列表",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/policy"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Policy"
        ],
        "summary": "修改policy",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/policy_update"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Policy"
        ],
        "summary": "删除policy",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/policy_delete"
              }
            }
          }
        }
      }
    },
    "/api/policys/addTags": {
      "post": {
        "tags": [
          "Policy"
        ],
        "summary": "给某一政策增加标签",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/policy_policyTags"
              }
            }
          }
        }
      }
    },
    "/api/policys/policyTags": {
      "post": {
        "tags": [
          "Policy"
        ],
        "summary": "？？？获取标签？？？",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/policy_policyTags"
              }
            }
          }
        }
      }
    },
    "/api/policys/deleteTags": {
      "post": {
        "tags": [
          "Policy"
        ],
        "summary": "给某一个政策删除Tag",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/policy_deleteTags"
              }
            }
          }
        }
      }
    },
    "/api/policys/add": {
      "post": {
        "tags": [
          "Policy"
        ],
        "summary": "插入policy",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/policy_add"
              }
            }
          }
        }
      }
    },
...
    "/api/questionnaires/createQuestionnaire": {
      "post": {
        "tags": [
          "Questionnaire"
        ],
        "summary": "创建questionnaire",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/recommendList_toggleNotice"
              }
            }
          }
        }
      }
    },
    "/api/questionnaires/getQuestionnaire": {
      "post": {
        "tags": [
          "Questionnaire"
        ],
        "summary": "获取questionnaire",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/questionnaire_get_update"
              }
            }
          }
        }
      }
    },
    "/api/questionnaires/updateQuestionnaire": {
      "post": {
        "tags": [
          "Questionnaire"
        ],
        "summary": "更新questionnaire",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/questionnaire_get_update"
              }
            }
          }
        }
      }
    },
    "/api/questionnaires/": {
      "delete": {
        "tags": [
          "Questionnaire"
        ],
        "summary": "删除questionnaire",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ]
      }
    }
  },
  "definitions": {
...
    "policy": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {}
        },
        "token": {
          "type": "string",
          "example": "4836f6e1-2e8b-4b45-b903-9cf88d990901"
        }
      }
    },
    "policy_add": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "POLICY_ID": {
              "type": "string",
              "example": ""
            },
            "POLICY_TITLE": {
              "type": "string",
              "example": "ddd"
            },
            "CDATA": {
              "type": "string",
              "example": "ddd"
            },
            "CREATED_AT": {
              "type": "string",
              "example": "2021-09-08T03:22:17.848Z"
            },
            "UPDATED_AT": {
              "type": "string",
              "example": "2021-09-08T03:22:17.848Z"
            },
            "POLICY_URL": {
              "type": "string",
              "example": " "
            },
            "USER_USER_ID": {
              "type": "string",
              "example": "2200"
            }
          }
        },
        "token": {
          "type": "string",
          "example": "4836f6e1-2e8b-4b45-b903-9cf88d990901"
        }
      }
    },
    "policy_update": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "POLICY_ID": {
              "type": "string",
              "example": "2cd0894c-0a25-11ec-a6b8-342eb7e2a2ec"
            },
            "POLICY_TITLE": {
              "type": "string",
              "example": "市政府关于印发《扬州市地方储备粮管理办法》的通知"
            },
            "CREATED_AT": {
              "type": "string",
              "example": "1970-01-01"
            },
            "UPDATED_AT": {
              "type": "string",
              "example": "1970-01-01"
            },
            "tags": {
              "type": "array",
              "example": [],
              "items": {}
            },
            "CDATA": {
              "type": "string",
              "example": "ddd"
            }
          }
        }
      }
    },
    "policy_delete": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "text": {
              "type": "string",
              "example": "是否要删除市政府关于印发《扬州市公共安全视频图像信息系统管理办法》的通知(ID: 2cd0894d-0a25-11ec-9af9-342eb7e2a2ec)?"
            },
            "id": {
              "type": "string",
              "example": "2cd0894d-0a25-11ec-9af9-342eb7e2a2ec"
            },
            "type": {
              "type": "string",
              "example": "policy"
            },
            "POLICY_ID": {
              "type": "string",
              "example": "2cd0894d-0a25-11ec-9af9-342eb7e2a2ec"
            }
          }
        }
      }
    },
    "policy_policyTags": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "POLICY_ID_POLICY_ID": {
              "type": "string",
              "example": "2cd03b48-0a25-11ec-96a4-342eb7e2a2ec"
            },
            "TAG_ID_TAG_ID": {
              "type": "string",
              "example": "3af9ddba-0a25-11ec-be60-90ccdf01187c"
            }
          }
        },
        "token": {
          "type": "string",
          "example": "4836f6e1-2e8b-4b45-b903-9cf88d990901"
        }
      }
    },
    "policy_deleteTags": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "POLICY_ID_POLICY_ID": {
              "type": "string",
              "example": "2cd03b48-0a25-11ec-96a4-342eb7e2a2ec"
            },
            "TAG_ID_TAG_ID": {
              "type": "string",
              "example": "3af9ddba-0a25-11ec-be60-90ccdf01187c"
            }
          }
        },
        "token": {
          "type": "string",
          "example": "4836f6e1-2e8b-4b45-b903-9cf88d990901"
        }
      }
    },
...
    "recommendList_toggleNotice": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "RECOMMENDED_LIST_ID": {
              "type": "number",
              "example": 111
            },
            "STATUS": {
              "type": "string",
              "example": ""
            }
          }
        }
      }
    },
...
    "questionnaire_get_update": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "ANSWERS": {
              "type": "object",
              "properties": {
                "TEST": {
                  "type": "string",
                  "example": "132"
                },
                "COOL": {
                  "type": "string",
                  "example": "it works"
                }
              }
            },
            "UUID": {
              "type": "string",
              "example": "455bdad0-4cbb-4bc4-85b0-0c34c0e4d8c4"
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "JiangduJWT": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      }
    }
  }
}